Updated dependencies and resolved most version conflicts (#1018)
